requirements.txt
google-cloud-vision
opencv-python
pillow
tkinter
numpy
